requirements.txt
typer
PyYAML
numpy
opencv-python
torch
torchvision
tqdm
Pillow
argparse 
hydra-core 
scipy
matplotlib
pandas
cognitive-face
filelock 
ffmpeg-python
scikit-learn
pyquaternion
decord